Fix op_type_id type mismatch: INT -> SMALLINT
HAF defines op_type_id as SMALLINT in hafd.operations and
hafd.operation_types tables, but HAfAH type definitions used INT.
This caused PostgREST to return HTTP 400:
"Returned type smallint does not match expected type integer in column 5."

Change hafah_backend.operation and hafah_backend.op_types to use
SMALLINT for op_type_id, matching the underlying HAF schema.

+++ backend/rest/operations/op_types.sql
@@ -37,7 +37,7 @@ AS
 $$
 BEGIN
   RETURN QUERY SELECT
-    id::INT,
+    id,
     -- Name parsing: Extract operation name from C++ fully-qualified name
     -- "hive::protocol::vote_operation" -> "vote_operation"
     split_part(name, '::', 3),
